refactor: merge ralimit class into client.py
